Started separating Cypress from node world. Reorganizing files.
