+++ lib/commands/common.ts
@@ -17,5 +17,6 @@ if (!Cypress.c8ypact) {
     config: {},
     getConfigValue: (key: string, defaultValue?: any) => undefined,
     getConfigValues: () => undefined,
+    loadCurrent: () => cy.wrap<C8yPact | null>(null, { log: false }),
   };
 }
